refactored RequestResultFactory to Factories namespace
